requirements.txt
requests~=2.31.0
python-dotenv~=0.21.0
psycopg2-binary~=2.9.10
transformers~=4.49.0
textblob~=0.19.0
torch
dbt-postgres